Next version of sliding

+++ src/renderer/plane_utils.cpp
@@ -290,7 +290,7 @@ std::pair<int, int> CreateIntersection(int r1_left, int r1_right, int r2_left, i
   }
 }
 
-double SegmentLength(int x1, int y1, int x2, int y2) {
+double SegmentLength(double x1, double y1, double x2, double y2) {
   return sqrt((x1 - x2)*(x1 - x2) + (y1 - y2)*(y1 - y2));
 }
